refactor(#77): 인벤토리 기본 정렬 로직 이동
