+++ src/widgets/inventory/ui/inventory-grid.tsx
@@ -1,7 +1,4 @@
-import type {
-  InventoryItem,
-  InventoryItemSlot,
-} from "@/entities/inventory/model/types";
+import type { InventoryItem } from "@/entities/inventory/model/types";
 import { InventoryItemCard } from "@/entities/inventory/ui/inventory-item-card";
 import { PixelSlotButton } from "@/shared/ui/pixel-slot-button";
 import { PixelCheckIcon } from "@/shared/ui/pixel-check-icon";
@@ -14,48 +11,23 @@ interface InventoryGridProps {
   items: InventoryItem[];
   selectedItemId?: string | null;
   onSelect: (item: InventoryItem) => void;
-  disableDefaultSort?: boolean;
 }
-
-const SLOT_ORDER: Record<InventoryItemSlot, number> = {
-  helmet: 0,
-  armor: 1,
-  weapon: 2,
-  ring: 3,
-  consumable: 4,
-};
 
 export function InventoryGrid({
   items,
   selectedItemId,
   onSelect,
-  disableDefaultSort = false,
 }: InventoryGridProps) {
   const { t } = useTranslation();
   const resolveItemName = useCatalogItemNameResolver();
-  const sortedItems = disableDefaultSort
-    ? items
-    : [...items].sort((a, b) => {
-        if (a.isEquipped !== b.isEquipped) {
-          return a.isEquipped ? -1 : 1;
-        }
-
-        if (SLOT_ORDER[a.slot] !== SLOT_ORDER[b.slot]) {
-          return SLOT_ORDER[a.slot] - SLOT_ORDER[b.slot];
-        }
-
-        return (
-          new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
-        );
-      });
 
   return (
     <PixelPanel title={t("inventory.grid.title")}>
-      {sortedItems.length === 0 ? (
+      {items.length === 0 ? (
         <PixelEmptyState message={t("inventory.empty")} />
       ) : (
         <div className="grid grid-cols-2 gap-3 sm:grid-cols-3 md:grid-cols-4 lg:grid-cols-5 xl:grid-cols-6">
-          {sortedItems.map((item) => (
+          {items.map((item) => (
             <InventoryGridCell
               key={item.id}
               item={item}
